Occupying new nodes while walking gives problem
MoveAbility now sets a node as occupied when entering it. The problem is that AStar now sees the target position as occupied when the unit is close to the goal, resulting in ... the unit going back and forth, back and forth between the end node and the closest node to that.